[scanner] refactor: split oversized missions components
Split 6 oversized missions components into smaller sibling files,
following the existing *.hooks.ts / *.parts.tsx / *.types.ts convention
used elsewhere in the codebase (e.g. FeedbackModal):

- MissionContentViewer.tsx: extracted useMissionContentViewer hook to
  MissionContentViewer.hooks.ts, and MissionFilePanel/MissionSearchPanel
  to MissionContentViewer.parts.tsx.
- MissionDetailView.tsx: extracted types/constants to
  MissionDetailView.types.ts, and CopyButton/StepCard/SectionBadge/
  extractCodeBlocks to MissionDetailView.parts.tsx.
- ResolutionHistoryPanel.tsx: extracted ResolutionCard and
  ResolutionDetailPanel to ResolutionHistoryPanel.parts.tsx.
- MissionLandingPage.tsx: extracted constants/tab defs to
  MissionLandingPage.constants.ts, slug-resolution fetch logic to
  MissionLandingPage.api.ts, and DashboardMockup/SectionBadge to
  MissionLandingPage.parts.tsx.
- SubmitToKBDialog.tsx: extracted CNCF project detection and
  console-kb format conversion helpers to SubmitToKBDialog.helpers.ts.
- StandaloneOrbitDialog.tsx: extracted ClusterScopeSection to
  StandaloneOrbitDialog.parts.tsx, and buildScopeString/constants to
  StandaloneOrbitDialog.helpers.ts.

Pure moves only — no markup, logic, or exported-API changes.

Fixes #21776, Fixes #21784, Fixes #21785, Fixes #21786, Fixes #21787
