+++ web/src/components/missions/MissionLandingPage.tsx
@@ -15,258 +15,18 @@
  */
 import { useState, useEffect } from 'react'
 import { useParams, useNavigate } from 'react-router-dom'
-import { validateMissionExport } from '../../lib/missions/types'
-import type { MissionExport, MissionStep } from '../../lib/missions/types'
+import type { MissionExport } from '../../lib/missions/types'
 import { getHomeBrowseMissionsRoute } from '../../config/routes'
 import { emitMissionError, emitPageView } from '../../lib/analytics'
-
-// ============================================================================
-// Constants
-// ============================================================================
-
-/** Timeout for fetching mission content from the API (ms) */
-const FETCH_TIMEOUT_MS = 10_000
-
-/** Maximum number of steps to preview before truncating */
-const MAX_PREVIEW_STEPS = 5
-
-/** Badge colors by mission type */
-const TYPE_COLORS: Record<string, string> = {
-  repair: 'bg-red-500/20 text-red-400 border-red-500/30',
-  troubleshoot: 'bg-orange-500/20 text-orange-400 border-orange-500/30',
-  deploy: 'bg-blue-500/20 text-blue-400 border-blue-500/30',
-  upgrade: 'bg-green-500/20 text-green-400 border-green-500/30',
-  analyze: 'bg-purple-500/20 text-purple-400 border-purple-500/30',
-  custom: 'bg-slate-500/20 text-slate-400 border-slate-500/30',
-}
-
-/** Default badge style for unknown types */
-const DEFAULT_TYPE_COLOR = 'bg-slate-500/20 text-slate-400 border-slate-500/30'
-
-/** Tab definitions for mission content sections */
-type TabId = 'install' | 'uninstall' | 'upgrade' | 'troubleshoot'
-
-interface TabDef {
-  id: TabId
-  label: string
-  icon: string
-  getSteps: (m: MissionExport) => MissionStep[]
-  emptyMessage: string
-}
-
-const TABS: TabDef[] = [
-  {
-    id: 'install',
-    label: 'Install',
-    icon: '📦',
-    getSteps: (m) => m.steps || [],
-    emptyMessage: 'Install steps not available.',
-  },
-  {
-    id: 'uninstall',
-    label: 'Uninstall',
-    icon: '🗑️',
-    getSteps: (m) => m.uninstall || [],
-    emptyMessage: 'Uninstall steps not yet available for this mission.',
-  },
-  {
-    id: 'upgrade',
-    label: 'Update / Upgrade',
-    icon: '⬆️',
-    getSteps: (m) => m.upgrade || [],
-    emptyMessage: 'Upgrade steps not yet available for this mission.',
-  },
-  {
-    id: 'troubleshoot',
-    label: 'Troubleshooting',
-    icon: '🔧',
-    getSteps: (m) => m.troubleshooting || [],
-    emptyMessage: 'Troubleshooting steps not yet available for this mission.',
-  },
-]
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-// ⚠️ PERFORMANCE CRITICAL — DO NOT CHANGE WITHOUT TESTING WITH CHROME CDP ⚠️
-//
-// This section uses smart prefix routing to resolve mission slugs to file
-// paths with 1-2 requests instead of the previous 13-directory brute-force.
-// The old approach fired 13 parallel requests (12 returning 404) and took
-// 10-20 seconds on cold cache. The current approach resolves in <2s.
-//
-// The MissionLandingPage route is also intentionally OUTSIDE the heavy
-// dashboard provider stack (see App.tsx LightweightShell) to avoid loading
-// 1.8MB of dashboard JS. Changing the route structure in App.tsx will
-// regress this. Always verify with:
-//   1. Clear browser cache via CDP: Network.clearBrowserCache
-//   2. Navigate to /missions/install-karmada
-//   3. Check: jsChunks < 20, totalJsKB < 300, apiCalls <= 3, pageLoadMs < 3000
-
-/**
- * Get the most likely file paths for a mission slug based on its prefix.
- * install-* → cncf-install/ or platform-install/
- * platform-* → platform-install/
- * Others → try slug as a subdirectory hint in cncf-generated/
- */
-function getPreferredPaths(slug: string): string[] {
-  if (slug.startsWith('install-')) {
-    return [
-      `fixes/cncf-install/${slug}.json`,
-      `fixes/platform-install/${slug}.json`,
-    ]
-  }
-  if (slug.startsWith('platform-')) {
-    return [`fixes/platform-install/${slug}.json`]
-  }
-  // For cncf-generated missions, the slug often starts with the project name
-  // e.g., "karmada-1234-some-issue" → cncf-generated/karmada/karmada-1234-some-issue.json
-  const projectHint = slug.split('-')[0]
-  return [
-    `fixes/cncf-generated/${projectHint}/${slug}.json`,
-    `fixes/security/${slug}.json`,
-    `fixes/troubleshoot/${slug}.json`,
-    `fixes/llm-d/${slug}.json`,
-    `fixes/multi-cluster/${slug}.json`,
-  ]
-}
-
-/**
- * Fetch a mission by slug. Tries the most likely paths first (1-2 requests),
- * then falls back to server-side slug resolution via index.json.
- */
-async function fetchMissionBySlug(slug: string): Promise<{ mission: MissionExport; raw: string } | null> {
-  // Fast path: try preferred directories based on slug prefix
-  for (const path of getPreferredPaths(slug)) {
-    try {
-      const url = `/api/missions/file?path=${encodeURIComponent(path)}`
-      const res = await fetch(url, { signal: AbortSignal.timeout(FETCH_TIMEOUT_MS) })
-      if (!res.ok) continue
-      const raw = await res.text()
-      const parsed = JSON.parse(raw)
-      const result = validateMissionExport(parsed)
-      if (result.valid) return { mission: result.data, raw }
-    } catch {
-      continue
-    }
-  }
-
-  // Fallback: search index.json for missions in unexpected directories
-  try {
-    const res = await fetch('/api/missions/file?path=fixes/index.json', {
-      signal: AbortSignal.timeout(FETCH_TIMEOUT_MS),
-    })
-    if (res.ok) {
-      const index = await res.json() as { missions?: Array<{ path: string }> }
-      const match = (index.missions || []).find((m) => {
-        const filename = (m.path || '').split('/').pop() || ''
-        return filename.replace('.json', '') === slug
-      })
-      if (match) {
-        const fileRes = await fetch(`/api/missions/file?path=${encodeURIComponent(match.path)}`, {
-          signal: AbortSignal.timeout(FETCH_TIMEOUT_MS),
-        })
-        if (fileRes.ok) {
-          const raw = await fileRes.text()
-          const parsed = JSON.parse(raw)
-          const result = validateMissionExport(parsed)
-          if (result.valid) return { mission: result.data, raw }
-        }
-      }
-    }
-  } catch {
-    // Fallback exhausted
-  }
-
-  return null
-}
-
-// ============================================================================
-// Dashboard Mockup Background
-// ============================================================================
-
-/** CSS-only mockup of the console dashboard — creates visual curiosity */
-function DashboardMockup() {
-  return (
-    <div className="absolute inset-0 overflow-hidden opacity-30 blur-[2px]">
-      {/* Sidebar */}
-      <div className="absolute left-0 top-0 bottom-0 w-[52px] bg-card border-r border-foreground/5">
-        {/* Logo area */}
-        <div className="h-12 flex items-center justify-center border-b border-foreground/5">
-          <div className="w-6 h-6 rounded bg-purple-500/30" />
-        </div>
-        {/* Nav items */}
-        {Array.from({ length: 8 }).map((_, i) => (
-          <div key={i} className="h-10 flex items-center justify-center">
-            <div className={`w-5 h-5 rounded ${i === 0 ? 'bg-purple-500/40' : 'bg-foreground/5'}`} />
-          </div>
-        ))}
-      </div>
-
-      {/* Main content area */}
-      <div className="ml-[52px] p-4">
-        {/* Top bar */}
-        <div className="h-10 mb-4 flex items-center gap-3">
-          <div className="w-32 h-6 rounded bg-foreground/5" />
-          <div className="flex-1" />
-          <div className="w-8 h-8 rounded-full bg-foreground/5" />
-          <div className="w-8 h-8 rounded-full bg-foreground/5" />
-        </div>
-
-        {/* Card grid */}
-        <div className="grid grid-cols-3 gap-3">
-          {/* Large card */}
-          <div className="col-span-2 h-48 rounded-xl bg-card border border-foreground/5 p-4">
-            <div className="w-24 h-3 rounded bg-foreground/8 mb-3" />
-            <div className="grid grid-cols-4 gap-2 h-32">
-              {Array.from({ length: 4 }).map((_, i) => (
-                <div key={i} className="rounded-lg bg-foreground/2 border border-foreground/5 p-2">
-                  <div className="w-full h-2 rounded bg-foreground/5 mb-2" />
-                  <div className="w-3/4 h-6 rounded bg-purple-500/10" />
-                </div>
-              ))}
-            </div>
-          </div>
-          {/* Tall card */}
-          <div className="row-span-2 rounded-xl bg-card border border-foreground/5 p-4">
-            <div className="w-20 h-3 rounded bg-foreground/8 mb-3" />
-            {Array.from({ length: 6 }).map((_, i) => (
-              <div key={i} className="flex items-center gap-2 mb-2">
-                <div className={`w-2 h-2 rounded-full ${i < 2 ? 'bg-red-500/40' : i < 4 ? 'bg-yellow-500/30' : 'bg-green-500/30'}`} />
-                <div className="flex-1 h-2 rounded bg-foreground/5" />
-              </div>
-            ))}
-          </div>
-          {/* Bottom row cards */}
-          <div className="h-40 rounded-xl bg-card border border-foreground/5 p-4">
-            <div className="w-16 h-3 rounded bg-foreground/8 mb-3" />
-            <div className="h-24 rounded bg-linear-to-t from-blue-500/5 to-transparent" />
-          </div>
-          <div className="h-40 rounded-xl bg-card border border-foreground/5 p-4">
-            <div className="w-20 h-3 rounded bg-foreground/8 mb-3" />
-            <div className="flex gap-1 h-24 items-end">
-              {Array.from({ length: 12 }).map((_, i) => (
-                <div
-                  key={i}
-                  className="flex-1 rounded-t bg-purple-500/15"
-                  style={{ height: `${20 + Math.sin(i * 0.8) * 40 + 30}%` }}
-                />
-              ))}
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Dim overlay to darken the mockup */}
-      <div className="absolute inset-0 bg-background/40" />
-    </div>
-  )
-}
-
-// ============================================================================
-// Component
-// ============================================================================
+import {
+  MAX_PREVIEW_STEPS,
+  TYPE_COLORS,
+  DEFAULT_TYPE_COLOR,
+  TABS,
+  type TabId,
+} from './MissionLandingPage.constants'
+import { fetchMissionBySlug } from './MissionLandingPage.api'
+import { DashboardMockup, SectionBadge } from './MissionLandingPage.parts'
 
 export function MissionLandingPage() {
   const { missionId } = useParams<{ missionId: string }>()
@@ -516,16 +276,3 @@ export function MissionLandingPage() {
   )
 }
 
-/** Small badge indicating whether a mission section has content */
-function SectionBadge({ present, label }: { present: boolean; label: string }) {
-  return (
-    <span className={`inline-flex items-center gap-1 px-2 py-0.5 text-2xs rounded-full border ${
-      present
-        ? 'text-green-400/70 bg-green-500/8 border-green-500/20'
-        : 'text-foreground/15 bg-foreground/2 border-foreground/5'
-    }`}>
-      <span className={`w-1.5 h-1.5 rounded-full ${present ? 'bg-green-500/60' : 'bg-foreground/10'}`} />
-      {label}
-    </span>
-  )
-}
